+++ protocols/perf/src/server.rs
@@ -44,7 +44,6 @@ impl Default for Behaviour {
                     request_response::ProtocolSupport::Inbound,
                 )),
                 request_response::Config::default()
-                    .with_connection_keep_alive(Duration::from_secs(60 * 5))
                     .with_request_timeout(Duration::from_secs(60 * 5)),
             ),
         }