+++ libp2p/CHANGELOG.md
@@ -4,7 +4,12 @@
   This supersedes the existing `libp2p::wasm_ext` module which is now deprecated.
   See [PR 3679].
 
+- Introduce a new `libp2p::SwarmBuilder` in favor of the now deprecated `libp2p::swarm::SwarmBuilder`.
+  See `libp2p::SwarmBuilder` docs on how to use the new builder.
+  Also see [PR 4120].
+
 [PR 3679]: https://github.com/libp2p/rust-libp2p/pull/3679
+[PR 4120]: https://github.com/libp2p/rust-libp2p/pull/4120
 
 ## 0.52.3